refactor(npc): Eliminados archivos antiguos y redireccionado correcto a los nuevos.
