--- src/roguelike_game/entities/npc/controllers/elite_controller.py
@@ -1,4 +1,0 @@
-# Path: src/roguelike_game/entities/npc/controllers/elite_controller.py
-from ..types.elite.controller import EliteController
-
-__all__ = ["EliteController"]
