--- src/roguelike_game/entities/npc/controllers/monster_controller.py
@@ -1,5 +1,0 @@
-#Path: src/roguelike_game/entities/npc/controllers/monster_controller.py
-
-from ..types.monster.controller import MonsterController
-
-__all__ = ["MonsterController"]
