--- src/roguelike_game/entities/npc/views/elite_view.py
@@ -1,4 +1,0 @@
-# Path: src/roguelike_game/entities/npc/views/elite_view.py
-from ..types.elite.view import EliteView
-
-__all__ = ["EliteView"]
