--- src/roguelike_game/entities/npc/views/monster_view.py
@@ -1,4 +1,0 @@
-# Path: src/roguelike_game/entities/npc/views/monster_view.py
-from ..types.monster.view import MonsterView
-
-__all__ = ["MonsterView"]
